Remove component and modify
