+++ package.json
@@ -15,6 +15,7 @@
     "@nextui-org/table": "^2.2.8",
     "@nextui-org/tabs": "^2.2.7",
     "@nextui-org/tooltip": "^2.2.7",
+    "framer-motion": "^11.16.1",
     "frontend-ocr-graduation": "file:",
     "next": "^15.1.3",
     "react": "^19.0.0",
